requirements.txt
# Core
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.0
python-dotenv==1.0.1

# LLM Integration
anthropic==0.39.0

# Data Processing
pandas==2.2.2
faker==28.4.1

# PII Detection & Masking
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355

# Database
sqlalchemy==2.0.35

# Utilities
rich==13.8.0
loguru==0.7.2
httpx==0.27.2
